requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
requests==2.32.3
pydantic==2.10.3
python-dotenv==1.0.1

SQLAlchemy==2.0.36
psycopg2-binary==2.9.11

tenacity==9.0.0
structlog==24.4.0

fhir.resources==8.0.0
pytest==8.3.4
